process_new_task function in the trait
